Feature/hours double (#119)
* Database and backend in double instead of int/

* Handle floats in the front-end

* Fixed tests

+++ src/main/kotlin/forms/HoliDayForm.kt
@@ -21,8 +21,8 @@ data class HoliDayForm(
     @JsonSerialize(using = LocalDateSerializer::class)
     override val to: LocalDate,
 
-    override val hours: Int,
-    override val days: List<Int>,
+    override val hours: Double,
+    override val days: List<Double>,
 
     override val status: Status = Status.REQUESTED,
